Feat: 삭제확인 모달 공용 컴포넌트 변환
